several optimizations.. no big gains.
Still having garbage collection issues with firefox. Two-to-three times in an animation, non-incremental GC is called and causes upwards of 300ms of jank. I can't figure out what's causing that as the profiler seems to be indicating random parts of the code. The only thing that I can think of is that due to the library being built with the revealing module pattern, it's causing too much to get stored on the heap, and then when it's time to call the render loop, GC gets forced.