ipad errata...
for some reason, some ios devices don't catch the promise error, but still fire a  event. When this happens, the video isn't paused, isn't stopped, but has a ready state of 0 and waits... for something?

This fix is more of a hack than anything else... I don't really understand whats happening.